requirements.txt
altgraph==0.17.3
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
attrs==23.1.0
bcrypt==4.1.2
bech32==1.2.0
bip-utils==2.9.0
blspy==2.0.2
btcpay==1.0.3
cachetools==5.3.1
cbor2==5.4.6
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
coincurve==17.0.0
cosmospy_protobuf==0.4.0
crcmod==1.7
cryptography==42.0.5
dnspython==2.3.0
docutils==0.19
ecdsa==0.18.0
ed25519-blake2b==1.4
exceptiongroup==1.2.0
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.62.1
h11==0.14.0
hdwallets==0.1.2
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib_metadata==7.0.2
jaraco.classes==3.3.1
jeepney==0.8.0
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
jwcrypto==1.5.6
keyring==24.3.1
keyrings.cryptfile==1.3.9
Kivy==2.2.1
Kivy-Garden==0.1.5
kivy-garden.mapview==1.0.6
kivymd==1.1.1
kivyoav==0.42
mapview==1.0.6
mnemonic==0.21
more-itertools==10.2.0
mospy-wallet==0.6.0
pexpect==4.8.0
Pillow==9.2.0
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
py-sr25519-bindings==0.2.0
pyarmor==8.5.12
pyarmor.cli.core==6.5.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydash==5.1.1
Pygments==2.13.0
pyinstaller==5.5
pyinstaller-hooks-contrib==2022.10
PyNaCl==1.5.0
pyparsing==3.1.0
python-dateutil==2.8.2
qrcode==7.3.1
referencing==0.29.1
requests==2.28.1
rpds-py==0.8.10
rsa==4.9
safe-pysha3==1.0.4
screeninfo==0.8.1
SecretStorage==3.3.3
sentinel-sdk==0.0.7
sentinel_protobuf==0.3.3
six==1.16.0
sniffio==1.3.1
stripe==4.2.0
toml==0.10.2
treelib==1.6.1
typing_extensions==4.10.0
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.18.1
